+++ src/test/groovy/net/fabricmc/loom/test/intergration/MultiProjectTest.groovy
@@ -22,21 +22,19 @@
  * SOFTWARE.
  */
 
-package net.fabricmc.loom
+package net.fabricmc.loom.test.intergration
 
-import com.google.common.hash.HashCode
-import com.google.common.hash.Hashing
-import com.google.common.io.Files
-import net.fabricmc.loom.util.ProjectTestTrait
+import net.fabricmc.loom.test.util.ArchiveAssertionsTrait
+import net.fabricmc.loom.test.util.ProjectTestTrait
 import spock.lang.Specification
 import spock.lang.Unroll
 
 import static org.gradle.testkit.runner.TaskOutcome.SUCCESS
 
-class ReproducibleBuildTest extends Specification implements ProjectTestTrait {
+class MultiProjectTest extends Specification implements ProjectTestTrait, ArchiveAssertionsTrait {
 	@Override
 	String name() {
-		"reproducible"
+		"multiproject"
 	}
 
 	@Unroll
@@ -45,20 +43,19 @@ class ReproducibleBuildTest extends Specification implements ProjectTestTrait {
 			def result = create("build", gradle)
 		then:
 			result.task(":build").outcome == SUCCESS
-			getOutputHash("fabric-example-mod-1.0.0.jar") == modHash
-			getOutputHash("fabric-example-mod-1.0.0-sources.jar") in sourceHash // Done for different line endings.
+			result.task(":core:build").outcome == SUCCESS
+			result.task(":example:build").outcome == SUCCESS
+
+			result.task(":remapAllJars").outcome == SUCCESS
+			result.task(":remapAllSources").outcome == SUCCESS
+
+			hasArchiveEntry("multiproject-1.0.0.jar", "META-INF/jars/example-1.0.0.jar")
+			hasArchiveEntry("multiproject-1.0.0.jar", "META-INF/jars/core-1.0.0.jar")
+			hasArchiveEntry("multiproject-1.0.0.jar", "META-INF/jars/fabric-api-base-0.2.1+9354966b7d.jar")
+
 		where:
-			gradle      | modHash                               | sourceHash
-			'6.8.3'     | "6132ffb4117adb7e258f663110552952"    | ["be31766e6cafbe4ae3bca9e35ba63169", "7348b0bd87d36d7ec6f3bca9c2b66062"]
-			'7.0-rc-1'  | "6132ffb4117adb7e258f663110552952"    | ["be31766e6cafbe4ae3bca9e35ba63169", "7348b0bd87d36d7ec6f3bca9c2b66062"]
-	}
-
-	String getOutputHash(String name) {
-		generateMD5(getOutputFile(name))
-	}
-
-	String generateMD5(File file) {
-		HashCode hash = Files.asByteSource(file).hash(Hashing.md5())
-		return hash.asBytes().encodeHex() as String
+			gradle              | _
+			DEFAULT_GRADLE      | _
+			PRE_RELEASE_GRADLE  | _
 	}
 }
