+++ src/test/groovy/net/fabricmc/loom/test/intergration/ReproducibleBuildTest.groovy
@@ -22,33 +22,43 @@
  * SOFTWARE.
  */
 
-package net.fabricmc.loom
+package net.fabricmc.loom.test.intergration
 
-import net.fabricmc.loom.util.ProjectTestTrait
+import com.google.common.hash.HashCode
+import com.google.common.hash.Hashing
+import com.google.common.io.Files
+import net.fabricmc.loom.test.util.ProjectTestTrait
 import spock.lang.Specification
 import spock.lang.Unroll
 
 import static org.gradle.testkit.runner.TaskOutcome.SUCCESS
 
-// This test runs a mod that exits on mod init
-class RunConfigTest extends Specification implements ProjectTestTrait {
+class ReproducibleBuildTest extends Specification implements ProjectTestTrait {
 	@Override
 	String name() {
-		"runconfigs"
+		"reproducible"
 	}
 
 	@Unroll
-	def "#task"() {
+	def "build (gradle #gradle)"() {
 		when:
-			def result = create(task)
+			def result = create("build", gradle)
 		then:
-			result.task(":${task}").outcome == SUCCESS
+			result.task(":build").outcome == SUCCESS
+			getOutputHash("fabric-example-mod-1.0.0.jar") == modHash
+			getOutputHash("fabric-example-mod-1.0.0-sources.jar") in sourceHash // Done for different line endings.
 		where:
-			task                | _
-			'runClient'         | _
-			'runServer'         | _
-			'runTestmodClient'  | _
-			'runTestmodServer'  | _
-			'runAutoTestServer' | _
+			gradle      | modHash                               | sourceHash
+			'6.8.3'     | "6132ffb4117adb7e258f663110552952"    | ["be31766e6cafbe4ae3bca9e35ba63169", "7348b0bd87d36d7ec6f3bca9c2b66062"]
+			'7.0-rc-1'  | "6132ffb4117adb7e258f663110552952"    | ["be31766e6cafbe4ae3bca9e35ba63169", "7348b0bd87d36d7ec6f3bca9c2b66062"]
+	}
+
+	String getOutputHash(String name) {
+		generateMD5(getOutputFile(name))
+	}
+
+	String generateMD5(File file) {
+		HashCode hash = Files.asByteSource(file).hash(Hashing.md5())
+		return hash.asBytes().encodeHex() as String
 	}
 }
