+++ src/test/groovy/net/fabricmc/loom/test/intergration/SimpleProjectTest.groovy
@@ -22,35 +22,42 @@
  * SOFTWARE.
  */
 
-package net.fabricmc.loom
+package net.fabricmc.loom.test.intergration
 
-import net.fabricmc.loom.util.ArchiveAssertionsTrait
-import net.fabricmc.loom.util.ProjectTestTrait
+import net.fabricmc.loom.test.util.ProjectTestTrait
 import spock.lang.Specification
 import spock.lang.Unroll
 
 import static org.gradle.testkit.runner.TaskOutcome.SUCCESS
 
-class AccessWidenerTest extends Specification implements ProjectTestTrait, ArchiveAssertionsTrait {
+class SimpleProjectTest extends Specification implements ProjectTestTrait {
 	@Override
 	String name() {
-		"accesswidener"
+		"simple"
 	}
 
 	@Unroll
-	def "accesswidener (gradle #gradle)"() {
+	def "build (gradle #gradle)"() {
 		when:
 			def result = create("build", gradle)
 		then:
 			result.task(":build").outcome == SUCCESS
-			getArchiveEntry("fabric-example-mod-1.0.0.jar", "modid.accesswidener") == expected().replaceAll('\r', '')
 		where:
-			gradle     | _
-			'6.8.3'    | _
-			'7.0-rc-1' | _
+			gradle              | _
+			DEFAULT_GRADLE      | _
+			PRE_RELEASE_GRADLE  | _
 	}
 
-	String expected() {
-		new File("src/test/resources/accesswidener/expected.accesswidener").text
+	@Unroll
+	def "#ide config generation"() {
+		when:
+			def result = create(ide)
+		then:
+			result.task(":${ide}").outcome == SUCCESS
+		where:
+			ide 		| _
+			'idea' 		| _
+			'eclipse'	| _
+			'vscode'	| _
 	}
 }
